+++ resourcemanager/src/main/java/com/resourcemanager/dao/ProjectDAO.java
@@ -9,6 +9,6 @@ public interface ProjectDAO {
 	public void addProject(Project p);
 	public void updateProject(Project p);
 	public List<Project> listProjects();
-	public Project getProjectById(int id);
-	public void removeProject(int id);
+	public Project getProjectById(Long id);
+	public void removeProject(Long id);
 }
