+++ resourcemanager/src/main/java/com/resourcemanager/dao/ResourceDAO.java
@@ -9,6 +9,6 @@ public interface ResourceDAO {
 	public void addResource(Resource p);
 	public void updateResource(Resource p);
 	public List<Resource> listResources();
-	public Resource getResourceById(int id);
-	public void removeResource(int id);
+	public Resource getResourceById(Long id);
+	public void removeResource(Long id);
 }
