+++ resourcemanager/src/main/java/com/resourcemanager/service/impl/ProjectServiceImpl.java
@@ -24,13 +24,13 @@ public class ProjectServiceImpl implements ProjectService {
 
 	@Override
 	@Transactional
-	public void deleteProject(int id) {
+	public void deleteProject(Long id) {
 		this.projectDAO.removeProject(id);
 	}
 
 	@Override
 	@Transactional
-	public Project getProjectById(int id) {
+	public Project getProjectById(Long id) {
 		return this.projectDAO.getProjectById(id);
 	}
